# Support widescreen per-layer storyboard masking