--- a/osu.Game/Storyboards/Drawables/DrawableStoryboardLayer.cs
+++ b/osu.Game/Storyboards/Drawables/DrawableStoryboardLayer.cs
@@ -8,7 +8,7 @@ using osu.Framework.Graphics.Containers;
 
 namespace osu.Game.Storyboards.Drawables
 {
-    public class DrawableStoryboardLayer : LifetimeManagementContainer
+    public class DrawableStoryboardLayer : CompositeDrawable
     {
         public StoryboardLayer Layer { get; }
         public bool Enabled;
@@ -23,17 +23,34 @@ namespace osu.Game.Storyboards.Drawables
             Origin = Anchor.Centre;
             Enabled = layer.VisibleWhenPassing;
             Masking = layer.Masking;
+
+            InternalChild = new LayerElementContainer(layer);
         }
 
-        [BackgroundDependencyLoader]
-        private void load(CancellationToken? cancellationToken)
+        private class LayerElementContainer : LifetimeManagementContainer
         {
-            foreach (var element in Layer.Elements)
-            {
-                cancellationToken?.ThrowIfCancellationRequested();
+            private readonly StoryboardLayer storyboardLayer;
 
-                if (element.IsDrawable)
-                    AddInternal(element.CreateDrawable());
+            public LayerElementContainer(StoryboardLayer layer)
+            {
+                storyboardLayer = layer;
+
+                Width = 640;
+                Height = 480;
+                Anchor = Anchor.Centre;
+                Origin = Anchor.Centre;
+            }
+
+            [BackgroundDependencyLoader]
+            private void load(CancellationToken? cancellationToken)
+            {
+                foreach (var element in storyboardLayer.Elements)
+                {
+                    cancellationToken?.ThrowIfCancellationRequested();
+
+                    if (element.IsDrawable)
+                        AddInternal(element.CreateDrawable());
+                }
             }
         }
     }