@peppy 2020-03-31
Not 100% sure about arbitrary 16:9 masking, as storyboards may have been made to support higher than that correctly (its not ever written in the spec as to what aspect ratio that setting means, right?)