@bdach 2020-03-31
No, I couldn't see any specific details on what the wide-screen storyboard setting did on the wiki, much less a mention of that aspect ratio. There are a few community resources and forum threads that mention this ratio, but I expect it has more to do with 16:9 becoming the *de facto* standard for screen size rather than with how the setting actually works.

As to whether there are beatmaps with storyboards designed for wider screens, I couldn't really say.