@peppy 2020-03-31
Can you include some reference storyboards which display *worse* without this? Just a few links would be useful in making a decision here.