@peppy 2020-03-31
Hmm, never mind. I didn't realise it was already implemented this way but just not masked correctly. Let's leave it as it is for now.